pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Michael Rodionov <rodionovspb9618@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-environ = "^0.10.0"
gunicorn = "^20.1.0"
drf-spectacular = "^0.26.1"
djangorestframework-simplejwt = "^5.2.2"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
marshmallow-dataclass = "^8.5.13"
pytest-django = "^4.5.2"
django-pytest = "^0.2.0"
pytest-factoryboy = "^2.5.1"
flake8 = "^6.0.0"
mypy = "^1.2.0"
django-stubs = { extras = ["compatible-mypy"], version = "^4.2.0" }
djangorestframework-stubs = { extras = ["compatible-mypy"], version = "^3.14.0" }

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = "True"

[tool.django-stubs]
django_settings_module = "todolist.settings"

[tool.django-stubs.extra_requirements]
drf = ["djangorestframework-stubs"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
